feat(hooks): 🔧 inline player factory into hook
